DRTVWR-418: Update with local copies of build-{open,wolf}ssl.bat
After the curl-7.54.1 vendor branch update, when I went to switch back to the
default branch, Mercurial told me I had untracked local copies of
curl/projects/build-{open,wolf}ssl.bat that would collide with newly-added
files with those same pathnames. I moved them aside, switched to default and
reinstated the local copies. These appear to be "better" than what just
arrived in the tarball in a few respects:

- '\n' instead of '\r\n' (though that may not be "better" for cmd.exe)
- newer copyright date
- fixed copyright download URL
- more sophisticated Perl detection logic: recovery logic added

Accordingly, I've checked in the files we had locally.

This may well be moot: I doubt we actually use either of these .bat files. But
from an information-preserving point of view, these have now been captured. If
it later turns out to be better to backout the changes, so be it.